requirements.txt
streamlit>=1.32.0
pytest>=8.0.0
playwright>=1.42.0
allure-pytest>=2.13.2
allure-commandline>=2.27.0
sqlalchemy>=2.0.0
pyyaml>=6.0.1
pandas>=2.2.0
watchdog>=4.0.0
python-dotenv>=1.0.1